added custom_size_based_buffer class
Added the feature for size and time based flushing of buffer.
Added config options for max_interval and max_size. Once either one is
reached the events stored in the buffer will be flushed.

Updated kusto.rb and ingestor.rb with the implementation of memory based
buffer and flushing and removed file based buffer