Added temporary file buffer for persistant storage
Updated upload_async and upload to raise errors during network downtime
to trigger rescue block in flush buffer so the data sent for flushing
can be restored and the flush can be attempted later.

Added file buffer to store data ONLY when flush fails due to network
issues. Once the network is back online the each file in the buffer is
flushed and deleted first and the regular in-memory buffer is used post
that.